reorganized navigation files into appropriate packages
